idp/requirements.txt
asyncpg~=0.29.0
fastapi~=0.111.0
uvicorn~=0.29.0
jwcrypto~=1.5.6
backoff~=2.2.1
urllib3==2.2.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-jose==3.3.0
aiohttp==3.9.5
